// src/app/model/menu-item.ts
export  class  MenuItem{

  name?: string;
  banglaName?: string;
  menuType?: number;
  menuTypeName?: string;
  serialNo?: number;
  url?: string;
  icon?: string;
  parentId?: number;
  reportUploadId?:number;
  reportUploadName?:string;
  insert?: boolean;
  edit?: boolean;
  delete?: boolean;
  view?: boolean;

}